+++ src/frontend/web_application/src/modules/draftMessage/components/DraftMessage/presenter.jsx
@@ -16,6 +16,7 @@ import { STATUS_DECRYPTED, STATUS_ERROR } from '../../../../store/modules/encryp
 import RecipientList from '../RecipientList';
 import AttachmentManager from '../AttachmentManager';
 import IdentitySelector from '../IdentitySelector';
+import RecipientSelector from '../RecipientSelector';
 import { getIdentityProtocol } from '../../services/getIdentityProtocol';
 
 import './draft-message-quick.scss';
@@ -201,8 +202,8 @@ class DraftMessage extends Component {
     return i18n._('draft-message.form.placeholder.quick-start', null, { defaults: 'Start a new discussion' });
   }
 
-  getIdentity = () => this.props.availableIdentities
-    .find(ident => ident.identity_id === this.state.draftMessage.identityId);
+  getIdentity = ({ identityId }) => this.props.availableIdentities
+    .find(ident => ident.identity_id === identityId);
 
   getCanSend = () => {
     const { isReply } = this.props;
@@ -226,7 +227,7 @@ class DraftMessage extends Component {
 
   validate = () => {
     const currentDraft = this.state.draftMessage;
-    const identity = this.getIdentity();
+    const identity = this.getIdentity({ identityId: this.state.draftMessage.identityId });
 
     if (!identity) {
       return [
@@ -386,13 +387,31 @@ class DraftMessage extends Component {
     });
   }
 
-  handleIdentityChange = ({ identity = {} }) => {
-    this.setState(prevState => ({
-      draftMessage: {
-        ...prevState.draftMessage,
-        identityId: identity.identity_id,
-      },
-    }), () => {
+  forceParticipantsProtocol = ({ protocol, participants }) => participants.map(participant => ({
+    ...participant,
+    protocol,
+  }));
+
+  handleIdentityChange = async ({ identity = {} }) => {
+    this.setState((prevState) => {
+      const { isReply } = this.props;
+      const prevIdentity = this.getIdentity({ identityId: prevState.draftMessage.identityId });
+      let recipients;
+
+      if (!isReply && prevIdentity && prevIdentity.protocol !== identity.protocol) {
+        recipients = this.forceParticipantsProtocol({
+          protocol: identity.protocol, participants: prevState.draftMessage.recipients,
+        });
+      }
+
+      return {
+        draftMessage: {
+          ...prevState.draftMessage,
+          ...(recipients ? { recipients } : {}),
+          identityId: identity.identity_id,
+        },
+      };
+    }, () => {
       const { internalId, original, onEditDraft } = this.props;
 
       return onEditDraft({
@@ -420,6 +439,19 @@ class DraftMessage extends Component {
         message: original,
       });
     });
+  }
+
+  handleChangeOne2OneRecipient = (ev) => {
+    // XXX: eventually select the identity that match the new protocol
+    const participant = ev.target.value;
+    this.setState(prevState => ({
+      draftMessage: {
+        ...prevState.draftMessage,
+        recipients: [
+          participant,
+        ],
+      },
+    }));
   }
 
   renderPlaceholder() {
@@ -458,11 +490,11 @@ class DraftMessage extends Component {
   }
 
   renderRecipientList({ className } = {}) {
-    const { canEditRecipients } = this.props;
+    const { canEditRecipients, isReply } = this.props;
 
     if (canEditRecipients) {
       const { internalId } = this.props;
-      const identity = this.getIdentity();
+      const identity = this.getIdentity({ identityId: this.state.draftMessage.identityId });
 
       return (
         <RecipientList
@@ -471,6 +503,23 @@ class DraftMessage extends Component {
           recipients={this.state.draftMessage.recipients}
           onRecipientsChange={this.handleRecipientsChange}
           identity={identity}
+        />
+      );
+    }
+
+    const isOne2One = isReply &&
+      this.state.draftMessage.recipients.length === 1 &&
+      this.state.draftMessage.recipients[0].contact_ids &&
+      this.state.draftMessage.recipients[0].contact_ids.length > 0;
+    const [recipient] = (isOne2One && this.state.draftMessage.recipients) || [];
+
+    if (isOne2One) {
+      return (
+        <RecipientSelector
+          className={className}
+          contactId={recipient.contact_ids[0]}
+          current={recipient}
+          onChange={this.handleChangeOne2OneRecipient}
         />
       );
     }
@@ -554,15 +603,14 @@ class DraftMessage extends Component {
     } = this.props;
 
     const encryptionEnabled = isEncrypted && encryptionStatus === STATUS_DECRYPTED;
-    const identity = this.getIdentity();
+    const identity = this.getIdentity({ identityId: this.state.draftMessage.identityId });
 
     const isSubjectSupported = ({ draft }) => {
       if (!draft.identityId) {
         return false;
       }
 
-      const currIdentity = availableIdentities
-        .find(ident => ident.identity_id === draft.identityId);
+      const currIdentity = this.getIdentity({ identityId: draft.identityId });
 
       if (!currIdentity) {
         return false;